storagecluster: Move all uninstall funcs to uninstall_reconciler.go
Move all uninstall related functions from reconcile.go to
uninstall_reconciler.go to organize them better
